refactor(agent-db): define the Ownership fence once, in one module
Cleanup pass over the Claim protocol; no behavior change, suite unchanged.

The fence rule was written twice in two dialects — once as raw SQL with
hand-written column names in `run-ownership.ts`, once rebuilt from Drizzle
builders inside lease renewal. A rename of any ownership column would have
type-checked and failed at runtime. It is now one predicate that renewal uses
directly and the `EXISTS` probe wraps, mirroring the
`ownedRunByUserConditions`/`ownedRunByUserExists` pair the Run side already
uses.

All of ADR-0015 moves into `conversation-ownership.ts`, leaving
`run-ownership.ts` untouched. The two authorities coexist only until the fenced
writes move onto the epoch, so a file boundary makes retiring the older one a
deletion rather than surgery inside a shared file.

The Claim's snapshot returns Run ids rather than rows: the array is already in
submission order, and no caller wanted the submission time it also carried.
The lease deadline expression is named, so "a renewal pushes it as far as the
Claim did" is structural instead of a comment.

Also corrects the `FenceRejection` doc, which claimed every fenced write
classifies into the vocabulary. Six writes still throw an opaque
`RunFenceError`; they adopt it as they are migrated.
