+++ AGENTS.md
@@ -144,7 +144,7 @@ chat-api holds no provider/document/E2B secrets. It admits Runs, attaches to AG-
 | `apps/agent-worker/src/artifacts/` | Successful-Run publication for Downloadable outputs: start/end workspace manifests, pre-upload `artifact_objects` ledgering, binary-safe private S3 upload, and the adapter seams used by PGlite tests |
 | `apps/agent-worker/src/sdk/` | SDK run supervision and production query wiring: `agent-stream.ts` emits standard Assistant text start/content, persists the canonical complete Assistant message, then emits text end; it also persists MyMemo-identified, correlated Tool lifecycle events before their live AG-UI projections while `running`, interrupts on abort, and fail-fast stops on `mirror_error` while reporting mirror reliability for supervisor reconciliation. The bound SessionStore supplies direct main-session mirror evidence, which `run-processor.ts` preserves as bounded turn metadata when the consumer reports a typed, otherwise-reliable stream failure so a concurrent durable interruption can reconcile continuity. The remaining tool/session/query/sandbox modules retain their documented ADR-0004/0005/0006/0012 responsibilities. `src/index.ts` wires the SDK processor and Redis Live Stream relay into the live worker |
 | `packages/live-text/` | Shared required Redis URL validation, producer-buffered in-memory/Redis relay implementations, event validation and bounds, and payload-safe delivery telemetry |
-| `packages/agent-db/` | Shared writable-DB data layer imported by chat-api + agent-worker: `src/schema.ts`, `src/client.ts` (`createDatabase`, `DbTx`), `src/run-ownership.ts` (Run-scoped and user-bound lease/ownership predicates, their `EXISTS` forms, `RunFenceError`, and the Conversation Ownership epoch fence probe of ADR-0015), `src/conversation-ownership.ts` (the Claim protocol: claim-plus-snapshot in one transaction, unconditional release, and epoch-fenced lease renewal — nothing executes through it yet), `src/run-store.ts` (queue helpers + terminal transaction composition, and the one typed `FenceRejection` vocabulary), `src/runtime-store.ts` (fenced sandbox/taint helpers, the named in-transaction `agent_session_id` update and recovery taint, and `orphan_sandboxes`), `src/session-store.ts` (worker-fenced SDK append/delete helpers plus deliberately unfenced transcript reads and administrative Conversation deletion over `agent_sessions`, ADR-0005), `src/artifact-store.ts` (pre-upload `artifact_objects` ledger and atomic pointer/current-metadata/`run_done` commit behind a fence-first Run lock), `src/run-events.ts` (`RunEventType` vocabulary), `src/testing.ts` (PGlite harness + `seedQueuedRun`), `src/migrations.ts` (`MIGRATIONS_DIR`), `drizzle/` migrations |
+| `packages/agent-db/` | Shared writable-DB data layer imported by chat-api + agent-worker: `src/schema.ts`, `src/client.ts` (`createDatabase`, `DbTx`), `src/run-ownership.ts` (Run-scoped and user-bound lease/ownership predicates, their `EXISTS` forms, and `RunFenceError`), `src/conversation-ownership.ts` (all of ADR-0015: the Ownership epoch fence and the Claim protocol — claim-plus-snapshot in one transaction, unconditional release, epoch-fenced lease renewal — with nothing executing through it yet), `src/run-store.ts` (queue helpers + terminal transaction composition, and the typed `FenceRejection` vocabulary), `src/runtime-store.ts` (fenced sandbox/taint helpers, the named in-transaction `agent_session_id` update and recovery taint, and `orphan_sandboxes`), `src/session-store.ts` (worker-fenced SDK append/delete helpers plus deliberately unfenced transcript reads and administrative Conversation deletion over `agent_sessions`, ADR-0005), `src/artifact-store.ts` (pre-upload `artifact_objects` ledger and atomic pointer/current-metadata/`run_done` commit behind a fence-first Run lock), `src/run-events.ts` (`RunEventType` vocabulary), `src/testing.ts` (PGlite harness + `seedQueuedRun`), `src/migrations.ts` (`MIGRATIONS_DIR`), `drizzle/` migrations |
 
 ### Chat Scopes
